*UPDATED* JdbiRequestRepositoryTests, clean build is working

+++ code/backend/LiftDropDomain/src/main/kotlin/pt/isel/liftdrop/Request.kt
@@ -3,10 +3,10 @@ package pt.isel.pipeline.pt.isel.liftdrop
 import pt.isel.liftdrop.RequestStatus
 
 data class Request(
-    val id: Long,
-    val clientId: Long,
-    val courierId: Long?,
+    val id: Int,
+    val clientId: Int,
+    val courierId: Int?,
     val requestStatus: RequestStatus,
-    val createdAt: Long,
+    val createdAt: Int,
     val details: RequestDetails,
 )
